+++ src/util/util.jl
@@ -8,11 +8,11 @@ function Base.convert(T::Type{Vector{SVector{N2,T1}}}, M::Matrix{T2}) where {N2,
     return Mout
 end
 
-function deleteat(M::Array{T,2}, i1::Int64, i2::Int64) where T
+function deleteat(M::Array{T,2}, i1::Int, i2::Int) where T
     [M[1:i1 - 1,1:i2 - 1] M[1:i1 - 1,i2 + 1:end];M[i1 + 1:end,1:i2 - 1] M[i1 + 1:end,i2 + 1:end]]
 end
 
-deleteat(M::Array{T,2},i::Int64) where T = deleteat(M, i, i)
+deleteat(M::Array{T,2},i::Int) where T = deleteat(M, i, i)
 
 skew(v::AbstractVector{T}) where T = SMatrix{3,3,T,9}(0, v[3], -v[2], -v[3], 0, v[1], v[2], -v[1], 0)
